+++ src/material.rs
@@ -12,7 +12,7 @@ use crate::vec3::{
 pub enum Material {
     Metal {
         albedo: Color,
-        fuzz: f64
+        fuzz: f32
     },
 
     Lambertian {
@@ -20,7 +20,7 @@ pub enum Material {
     },
 
     Dielectric {
-        index_of_refraction: f64
+        index_of_refraction: f32
     }
 }
 
@@ -57,13 +57,13 @@ impl Material {
                 let unit_direction = unit_vector(r_in.direction());
 
                 let dot_product = dot(-unit_direction, rec.normal);
-                let cos_theta: f64 = if dot_product < 1.0 {
+                let cos_theta: f32 = if dot_product < 1.0 {
                     dot_product
                 } else {
                     1.0
                 };
 
-                let sin_theta: f64 = (1.0 - cos_theta * cos_theta).sqrt();
+                let sin_theta: f32 = (1.0 - cos_theta * cos_theta).sqrt();
 
                 let cannot_refract: bool = refraction_ratio * sin_theta > 1.0;
                 let direction: Vec3;
@@ -82,9 +82,9 @@ impl Material {
 }
 
 
-fn reflectance(cosine: f64, ref_idx: f64) -> f64 {
+fn reflectance(cosine: f32, ref_idx: f32) -> f32 {
     // use Schlick's Approximation
-    let mut r0: f64 = (1.0 - ref_idx)  / (1.0 + ref_idx);
+    let mut r0: f32 = (1.0 - ref_idx)  / (1.0 + ref_idx);
     r0 *= r0;
     r0 + (1.0 - r0) * (1.0 - cosine).powf(5.0)
 }
