+++ src/sphere.rs
@@ -6,12 +6,12 @@ use crate::vec3::{Point3, Vec3};
 
 pub struct Sphere {
     center: Point3,
-    radius: f64,
+    radius: f32,
     material: Material,
 }
 
 impl Sphere {
-    pub fn new(center: Point3, radius: f64, material: Material) -> Self {
+    pub fn new(center: Point3, radius: f32, material: Material) -> Self {
         Sphere {
             center: center,
             radius: radius,
@@ -21,7 +21,7 @@ impl Sphere {
 }
 
 impl Hittable for Sphere {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = r.origin() - self.center;
         let a = r.direction().length_squared();
         let half_b = dot(oc, r.direction());
