+++ src/vec3.rs
@@ -8,46 +8,46 @@ use crate::utility::{random_double, random_double_range};
 
 #[derive(Debug, PartialEq, Copy, Clone)]
 pub struct Vec3 {
-    pub elements: [f64; 3]
+    pub elements: [f32; 3]
 }
 
 pub type Point3 = Vec3;
 pub type Color = Vec3;
 
 impl Vec3 {
-    pub fn new(e0: f64, e1: f64, e2: f64) -> Vec3 {
+    pub fn new(e0: f32, e1: f32, e2: f32) -> Vec3 {
         Vec3 {elements: [e0, e1, e2]}
     }
 
-    pub fn x(&self) -> f64 {
+    pub fn x(&self) -> f32 {
         self.elements[0]
     }
 
-    pub fn y(&self) -> f64 {
+    pub fn y(&self) -> f32 {
         self.elements[1]
     }
 
-    pub fn z(&self) -> f64 {
+    pub fn z(&self) -> f32 {
         self.elements[2]
     }
 
-    pub fn r(&self) -> f64 {
+    pub fn r(&self) -> f32 {
         self.elements[0]
     }
 
-    pub fn g(&self) -> f64 {
+    pub fn g(&self) -> f32 {
         self.elements[1]
     }
 
-    pub fn b(&self) -> f64 {
+    pub fn b(&self) -> f32 {
         self.elements[2]
     }
 
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         (self.length_squared()).sqrt()
     }
 
-    pub fn length_squared(&self) -> f64 {
+    pub fn length_squared(&self) -> f32 {
         self.x() * self.x() + self.y() * self.y() + self.z() * self.z()
     }
 
@@ -66,21 +66,21 @@ impl Vec3 {
         }
     }
 
-    pub fn random_range(min: f64, max: f64) -> Vec3 {
+    pub fn random_range(min: f32, max: f32) -> Vec3 {
         Vec3 {
             elements: [random_double_range(min, max), random_double_range(min, max), random_double_range(min, max)]
         }
     }
 
     pub fn near_zero(&self) -> bool {
-        let s: f64 = 1e-8;
+        let s: f32 = 1e-8;
 
         (self.x().abs() < s) && (self.y().abs() < s) && (self.z().abs() < s)
     }
 
 }
 
-pub fn dot(v1: Vec3, v2: Vec3) -> f64 {
+pub fn dot(v1: Vec3, v2: Vec3) -> f32 {
     v1.x() * v2.x() +
         v1.y() * v2.y() +
         v1.z() * v2.z()
@@ -125,7 +125,7 @@ pub fn reflect(v: Vec3, n: Vec3) -> Vec3 {
     v - 2.0 * dot(v, n) * n
 }
 
-pub fn refract(uv: Vec3, n: Vec3, etai_over_etat: f64) -> Vec3 {
+pub fn refract(uv: Vec3, n: Vec3, etai_over_etat: f32) -> Vec3 {
     let dot_product = dot(-uv, n);
     let cos_theta = if dot_product < 1.0 {
         dot_product
@@ -239,10 +239,10 @@ impl MulAssign<Vec3> for Vec3 {
     }
 }
 
-impl Mul<f64> for Vec3 {
+impl Mul<f32> for Vec3 {
     type Output = Self;
 
-    fn mul(self, value: f64) -> Self {
+    fn mul(self, value: f32) -> Self {
         Self {
             elements: [
                 self.x() * value,
@@ -253,7 +253,7 @@ impl Mul<f64> for Vec3 {
     }
 }
 
-impl Mul<Vec3> for f64 {
+impl Mul<Vec3> for f32 {
     type Output = Vec3;
 
     fn mul(self, vec3: Vec3) -> Vec3 {
@@ -275,15 +275,15 @@ impl Div<Vec3> for Vec3 {
     }
 }
 
-impl Div<f64> for Vec3 {
+impl Div<f32> for Vec3 {
     type Output = Self;
 
-    fn div(self, value: f64) -> Self {
+    fn div(self, value: f32) -> Self {
         (1.0 / value) * self
     }
 }
 
-impl Div<Vec3> for f64 {
+impl Div<Vec3> for f32 {
     type Output = Vec3;
 
     fn div(self, vec3: Vec3) -> Vec3 {
@@ -383,7 +383,7 @@ mod tests {
     #[test]
     fn vec3_mul_value_1() {
         let mut test_vec = Vec3::new(1.0, 2.0, 3.0);
-        let value: f64 = 2.0;
+        let value: f32 = 2.0;
         let product = Vec3::new(2.0, 4.0, 6.0);
         
         assert_eq!(test_vec * value, product);
@@ -392,7 +392,7 @@ mod tests {
     #[test]
     fn vec3_mul_value_2() {
         let mut test_vec = Vec3::new(1.0, 2.0, 3.0);
-        let value: f64 = 2.0;
+        let value: f32 = 2.0;
         let product = Vec3::new(2.0, 4.0, 6.0);
         
         assert_eq!(value * test_vec, product);
@@ -410,7 +410,7 @@ mod tests {
     #[test]
     fn vec3_div_value_1() {
         let test_vec = Vec3::new(1.0, 2.0, 3.0);
-        let value: f64 = 2.0;
+        let value: f32 = 2.0;
         let quotient = Vec3::new(0.5, 1.0, 1.5);
 
         assert_eq!(test_vec / value, quotient);
@@ -419,7 +419,7 @@ mod tests {
     #[test]
     fn vec3_div_value_2() {
         let test_vec = Vec3::new(1.0, 2.0, 3.0);
-        let value: f64 = 2.0;
+        let value: f32 = 2.0;
         let quotient = Vec3::new(0.5, 1.0, 1.5);
 
         assert_eq!(value / test_vec, quotient);
